Performance upgrades and additional improvements
Split the suggestion_algorithm function into two parts. The part that only needs to be completed on the first pass and the part that needs to be completed on evey pass.

Removed the need for having two global lists, likes_to_update and likes_to_save. Now only has the local variable likes_to_save.

Made the list of classes global because previously it was being passed into every function.

Functions were previously returning values when they were not needed. This has been addressed in this update.